TQAunit mistake container was extended to 32 bit to support more kinds of mistakes. All dependices adjusted to it. Reading 8bit from old binary exam files is still supported.

+++ src/exam/texecutorsupply.cpp
@@ -88,8 +88,8 @@ QColor& TexecutorSupply::answerColor(const TQAunit& answer) {
 }
 
 
-QColor& TexecutorSupply::answerColor(quint8 mistake) {
-	if (mistake == (quint8)TQAunit::e_correct)
+QColor& TexecutorSupply::answerColor(quint32 mistake) {
+	if (mistake == (quint32)TQAunit::e_correct)
     return gl->EanswerColor;
   else if (!(mistake & 16) && !(mistake & 64))
     return gl->EnotBadColor;
